Creating MeetingInvites style file
