+++ src/Styles/RegisterStyles.js
@@ -63,6 +63,6 @@ const styles = StyleSheet.create({
       fontSize: 18,
     },
   
-  });
-  
-  export default styles;
+});
+
+export default styles;
